third_party/blink/web_tests/external/wpt/css/selectors/is-specificity-shadow-expected.txt

This is a testharness.js-based test.
FAIL Specificity of :host() inside :is() assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
